Move msgpack encoding into separate package
